+++ lib/ui_icons.ex
@@ -7,14 +7,14 @@ defmodule PrimerLive.UIIcons do
 
   require PrimerLive.Helpers.DeclarationHelpers
 
-  alias PrimerLive.Helpers.{AttributeHelpers, DeclarationHelpers}
-
   @common_svg_attrs [
     viewBox: "0 0 16 16",
     xmlns: "http://www.w3.org/2000/svg",
     "aria-hidden": "true",
     focusable: "false"
   ]
+
+  def common_svg_attrs, do: @common_svg_attrs
 
   @doc """
   Internal use: generates a map of SVG's.
@@ -94,54 +94,5 @@ defmodule PrimerLive.UIIcons do
       </svg>
       """
     }
-  end
-
-  attr(:size, :any, default: 32, doc: "Spinner size (number or number as string).")
-  attr(:color, :string, default: "currentColor", doc: "Spinner color as SVG fill color.")
-  attr(:gap_color, :string, default: "currentColor", doc: "Spinner gap color as SVG fill color.")
-  DeclarationHelpers.class()
-  DeclarationHelpers.rest()
-
-  def ui_icon_spinner(assigns) do
-    class =
-      AttributeHelpers.classnames([
-        "anim-rotate",
-        assigns[:class]
-      ])
-
-    svg_attrs =
-      AttributeHelpers.append_attributes(assigns.rest, [
-        [class: class],
-        [width: assigns.size],
-        [height: assigns.size]
-      ])
-
-    assigns =
-      assigns
-      |> assign(:common_attrs, @common_svg_attrs)
-      |> assign(:svg_attrs, svg_attrs)
-
-    ~H"""
-    <svg {@common_attrs} fill="none" {@svg_attrs} {@rest}>
-      <circle
-        cx="8"
-        cy="8"
-        r="7"
-        stroke={@color}
-        stroke-opacity="0.25"
-        stroke-width="2"
-        vector-effect="non-scaling-stroke"
-      >
-      </circle>
-      <path
-        d="M15 8a7.002 7.002 0 00-7-7"
-        stroke={@gap_color}
-        stroke-width="2"
-        stroke-linecap="round"
-        vector-effect="non-scaling-stroke"
-      >
-      </path>
-    </svg>
-    """
   end
 end