+++ test/components/spinner_test.exs
@@ -56,7 +56,26 @@ defmodule PrimerLive.Components.SpinnerTest do
              |> format_html()
   end
 
-  test "Attribute: gap color" do
+  test "Attribute: highlight_color" do
+    assigns = %{}
+
+    assert rendered_to_string(~H"""
+           <.spinner highlight_color="black" />
+           <.spinner highlight_color="#000000" />
+           <.spinner highlight_color="rgba(0, 0, 0, 1)" />
+           """)
+           |> format_html() ==
+             """
+             <svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none"
+             class="anim-rotate" height="32" width="32">STRIPPED_SVG_PATHS</svg><svg viewBox="0 0 16 16"
+             xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none" class="anim-rotate" height="32"
+             width="32">STRIPPED_SVG_PATHS</svg><svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" aria-hidden="true"
+             focusable="false" fill="none" class="anim-rotate" height="32" width="32">STRIPPED_SVG_PATHS</svg>
+             """
+             |> format_html()
+  end
+
+  test "Attribute: gap_color (deprecated)" do
     assigns = %{}
 
     assert rendered_to_string(~H"""